Log User Sign In Session

+++ app/components/form.tsx
@@ -26,23 +26,30 @@ export default function AuthForm() {
 
         client.read({ limit: 99999 }).then(function (data: any) {
           let latestUsers = JSON.parse(data);
+          let existingUsers = latestUsers.filter((usr: any) => usr?.email == email);
           let potentialUser = { id: latestUsers.length + 1, name: name, email: email, password: password, updated: updated, lastSignin: lastSignin, registered: registered, roles: [`user`] };
-
-          client.create(potentialUser).then(function (data: any) {
-            setUser(potentialUser);
-            localStorage.setItem(`user`, JSON.stringify(potentialUser));
+          
+          if (existingUsers.length == 1) {
+            setUser(existingUsers[0]);
+            localStorage.setItem(`user`, JSON.stringify(existingUsers[0]));
             setUpdates(updates + 1);
-            client.read({ limit: 99999 }).then(function (data: any) {
-              let databaseData = JSON.parse(data);
-              console.log(`Updated Users`, databaseData, `with Data`, potentialUser);
+            updateUser(`id`, existingUsers[0]?.id, { updated: formatDate(new Date()) });
+          } else {
+            client.create(potentialUser).then(function (data: any) {
+              setUser(potentialUser);
+              localStorage.setItem(`user`, JSON.stringify(potentialUser));
+              setUpdates(updates + 1);
+              client.read({ limit: 99999 }).then(function (data: any) {
+                let databaseData = JSON.parse(data);
+                console.log(`Updated Users`, databaseData, `with Data`, potentialUser);
+              });
+            }, function (err: any) {
+              console.log(err);
             });
-          }, function (err: any) {
-            console.log(err);
-          });
+          }
         }, function (err: any) {
           console.log(err);
         });
-        // updateUser(`id`, 1, { updated: formatDate(new Date()) });
       } else {
         setUser(null);
         setUpdates(updates+1);
@@ -52,7 +59,7 @@ export default function AuthForm() {
 
   const updateUser = (fieldName: any, fieldValue: any, valueToUpdateObj: any) => {
     client.update(fieldName, fieldValue, valueToUpdateObj).then(function (data: any) {
-      console.log(data);
+      return;
     });
   }
